Rework the RDA_isr interrupt code
The RDA interrupt routine handles reading data from the serial device
into a buffer.  The code is reworked primarily to avoid calling
getc() when there is nothing to read.

There is still improvements to be made here.  The deficienies are largly
called out in the code comments.

First of all because this is a psuedo-circular buffer, where there always
needs to be enough space in the buffer to hold one line, lines larger
than MAX_SAFE_LINE_SIZE will be split into two if they happen to be read
when the buffer is nearly full.  This was an issue in the original code
as well, we just make it more clear that this is happening.  This could
be solved if we rework the code to use a true byte-oriented circular buffer.
But that would require a larger overhaul of the new process_line_from_pc()
function.

Next, if the RDA_isr routine catches up with the point where main() is
consuming the data, it now pauses.  Previously there was a potential message
corruption point here.  So this is somewhat of an improvement, expect that
reading will pause until the next interrupt occurs.  If no interrupts are
forthcoming, data will sit unread.  If the RDA interrupt can be raised
in software on the PIC, we could raise the interrupt from
buf_out_ptr_advance().

Alternatively, we could make the contents of RDA_isr() a separate function that
can be called from both RDA_isr() and buf_out_ptr_advance() (it would be called
in the latter while the interrupt is disabled).  If we did that, we might get
really fancy and have RDA_isr() set a flag when it stops processing to let
buf_out_ptr_advance() know that it has to read the pending data itself.

Another future enhancement: have process_line_from_pc() return the lenght of
the processed line (it needs to walk the whole line anyway), so that we can
pass that value to buf_out_ptr_advance() and avoid the redundant strlen()
call in there.